Add files via upload
Alright. ACSModeSequence is how we input the commanded values for each block. Constants_gui.m is the same as before, but I incorporated calling the function ACSModeSequence.m. Also note that ACSModeSequence.m has all the constant blocks named after SimulatorSpencer. If we change this, we will need to change these constant block names. There is a comment that you can just comment out the function call and work as before. I have uploaded SimulatorSpencer as well also as this has changes from what Rob uploaded yesterday, notably a removal of the extra nodes and change of the GUI interface. The GUI interface also has Real Time Synchronization block. If you do not have Windows, just comment this out and it should run. Note that many errors pop up, but it runs relatively ok. Press emergency stop if beginning to stall so that you do not need to Task Manager close the entire thing. If any questions please ask me.